chore: Resolve conflicts and bring the branch up-to-date with development
